requirements.txt
Flask==3.1.0
python-dotenv==1.0.1
pymongo==4.11.1
